+++ buss-20220822/ChangeLog.md
@@ -1,3 +1,14 @@
+2025-05-07 Version: 1.1.0
+- Support API GetSecurityEventDetail.
+- Support API RecordClickLinkAction.
+- Support API UpdateSecurityEventStatus.
+- Update API FindInstanceInfo: add request parameters businessCodes.
+- Update API SearchPunishRequest: add response parameters Body.DataList.$.Operator.
+- Update API SearchPunishRequest: add response parameters Body.DataList.$.OperatorNum.
+- Update API SearchPunishRequest: add response parameters Body.DataList.$.PunishOfficer.
+- Update API SearchPunishRequest: add response parameters Body.DataList.$.PunishOfficerNum.
+
+
 2024-07-12 Version: 1.0.0
 - Generated python 2022-08-22 for Buss.